ci: fix gcc conversion issue
../src/handmadehero.c: In function ‘LoadBmp’:
../src/handmadehero.c:237:17: error: conversion from ‘u32’ {aka ‘unsigned int’} to ‘f32’ {aka ‘float’} may change value [-Werror=conversion]
  237 |         f32 a = (value >> alphaShift) & 0xff;
      |                 ^
../src/handmadehero.c:238:17: error: conversion from ‘u32’ {aka ‘unsigned int’} to ‘f32’ {aka ‘float’} may change value [-Werror=conversion]
  238 |         f32 r = (value >> redShift) & 0xff;
      |                 ^
../src/handmadehero.c:239:17: error: conversion from ‘u32’ {aka ‘unsigned int’} to ‘f32’ {aka ‘float’} may change value [-Werror=conversion]
  239 |         f32 g = (value >> greenShift) & 0xff;
      |                 ^
../src/handmadehero.c:240:17: error: conversion from ‘u32’ {aka ‘unsigned int’} to ‘f32’ {aka ‘float’} may change value [-Werror=conversion]
  240 |         f32 b = (value >> blueShift) & 0xff;

+++ src/handmadehero.c
@@ -234,10 +234,10 @@ LoadBmp(pfnPlatformReadEntireFile PlatformReadEntireFile, char *filename)
         u32 value = *srcDest;
 
         // extract pixel from file
-        f32 a = (value >> alphaShift) & 0xff;
-        f32 r = (value >> redShift) & 0xff;
-        f32 g = (value >> greenShift) & 0xff;
-        f32 b = (value >> blueShift) & 0xff;
+        f32 a = (f32)((value >> alphaShift) & 0xff);
+        f32 r = (f32)((value >> redShift) & 0xff);
+        f32 g = (f32)((value >> greenShift) & 0xff);
+        f32 b = (f32)((value >> blueShift) & 0xff);
 
         /*
          * Store channels values pre-multiplied with alpha.
